Basic Room Changes
I have resized the room to fit better on the screen, and I have also
placed some blocks down to form a floor. For the moment I am not sure
that it is important to have several separated areas, as we can just use
this one platform to test out mechanics for our different objects.