Sprite Creation & Gravity Implementation
I have stolen the sprite from the game, Braid, to use temporarily until
we get some actual art from the game. There is a sprite object, and both
the sprite object and the wall object interact using gravity generated
by the game. There are two other objects called "oStaticParent" and
"oDynamicParent" that are used to specify what moves and what doesn't.
To make something solid you add it as a child of the static object, and
vice versa if you want something to be moved by gravity.