Move Left/Right & Gravity Modifications
Gravity worked well, but it meant that we could not move the player side
to side without creating a sideways gravity variable that seemed
counter-intuitive. Instead, the player moves down until it contacts the
floor. I removed the two gravity objects, and set it up so that the
player moves left and right using the arrow keys, very simply for the
time being.