Left Sprite & Floor Clipping Fix
Player sprite is now mirrored when moving left. The issue with the
player clipping into the floor was caused by the fact that each step was
not a uniform length, as the gravity added 0.8 to the vspeed each time a
step was called. The player was therefore moved further than the
distance to the ground before the step detected that it was in the wrong
place. I tweaked the code a bit so that it detects the next step and
stops before moving. Then, it moves the sprite to meet the floor.