Horizontal Velocity for the Player
I have built this into the project, but I am sure that it can be
improved on. I will leave this issue open for the time being but the
animation will probably look better once we have a sprite map so that
our player doesn't look like he is ice-skating all of the time.