Standing Sprite with Animation & Finite State Machines Implementation
There are a number of changes in this push. I have added another sprite
with an animation sequence of the player standing still, which looks a
lot better due to the subtle movement. At the moment it is a little
broken due to the fact that the sprites are not uniform in size, so when
the sprites transition between each other it looks a bit weird. I will
get around to fixing this, but it took me quite a while to separate each
of the sprites and put them into the proper location so I will likely do
it late tomorrow. Until then, it shouldn't affect any other aspects of
the game. I have also begun to implement the game mechanics that you
mentioned, though I have not done a lot with them. Currently, everything
is still in the same file, however there is an enum in the first level
that references two script files in order to change between the two
different sprites.