Left Bug Fix, Jump Tweak, and Animation Enhancement
I fixed the bug with the moving left. There were a few issues with the
code, but the main one was that we had an if statement checking if the
left button was pressed, but the value of left was -1 instead of 1 so it
did not equivalate to true. I also changed the keycheck for jump so that
it does not queue up a jump event if you press jump in mid air. As a
nice side effect, you can now hold down the space bar to continually
jump. I also changed a couple of if statements to include "xor"s so that
we can have the player standing rather than running on the spot if the
player presses both left and right at once.