Jump/Fall Animation, Move in the Air, Code Cleanup
There are quite a few additions here, the main being that there is now a
jump animation. If the player jumps further than there are frames in the
animation, it transitions into a looping fall animation. To test the
animation, I have added a bunch of blocks to the room in order to add
height to the jump. I have also implemented a basic collision detected,
but it is pretty bad because I only needed it to work on a basic level.
It is now possible to move in mid air, as we had an issue with the way
that our loops were nested. Finally, I have removed any excess comments
as they are obsolete due to our functioning state system.