Esc Key Bug Fix & Moving between rooms
I changed the Esc detection code slightly as if you pressed it between
frames it did not register properly. I have additionally set it up so
that you can move back into the original room by pressing Esc again.
Currently it restarts the room which is useful for testing our
animations and such, but I will work on the ability to have it resume
where you left off.