Fall Animation & Restart Level With Main Menu
I have added a fall animation that triggers when the player walks off of
a surface. It is a little buggy in the way that it detects the surface
below it though, see the issue on GitHub for details. I have also added
the ability to restart the game without closing the window and rerunning
it. You can do this by clicking on one of the buttons in the Esc menu.
If you press Esc again, the level continues where you left it. This will
work for now and is mainly for bug testing, as I am aware that all of
the current buttons follow this behaviour, and it will always take you
back to the first room, no matter which room you pause from. For these
reasons, we will need to find a better solution in the future.