Allow player to stop moving block while holding shift.
The player sprite continues to run while doing this which is strange,
but I will not try to fix this until after everything else is done. I
tried to quickly but it broke a lot of other stuff so we will see. In
addition, if you press the opposite direction the block continues to
move in the initial direction.